clean up and abstract logic into single models file and single builder file. remove the reader cli - it will never be used.
